Update main.go
A DDF based on a static yield function: y=425000*x^1.5 where x = PoSreward / ticketprice.

It illustrates how the price declines as the PoSreward goes down over time. There is still a resonance buying behaviour. This is due to the cyclical nature of the algorithm and the fraction of buying power when the yieldTickets function is lower than 1. Due to it being fraction based, however, it does not jump insanely high, price only deviates by about 10, declining to 5.