Make surge/drop points relative to num blocks.
This modifies the code which simulates a sudden surge and drop in the
staked coins from 40% of the supply to 50% of the supply to occur at
points relative to the number of blocks being simulated as opposed to
hard coded values.

In particular, it uses 60% and 80%.  Thus, if 100,000 blocks are being
simulated, the surge will happen at block 60k and the drop at 80k.