Use DDF before stake validation height as well.
This modifies the simulator to use the DDF prior to the stake validation
height instead of just assuming 50% demand.  While the 50% more closely
matched what happened on mainnet, it was artificially limited by the
fact the system was new and relatively difficult to use at the time as
opposed to rational behavior.