Add highlight to surge region and amplify surge.
This modifies the generated graph to include a highlight in the section
where a surge of coins is introduced and then removed.  In addition, it
amplifies the surge by moving from between 40-50% stake participation to
40-60% stake participation.

While here, also condense the output of the results a bit by combining
like fields and add some usage notes for the graphs.