Double the demand during surge interval.
This amplifies the surges a little more by also doubling the demand
during that period.  This is more rational because it stands to reason
that a large portion of coins suddenly wanting to enter the staking
system would increase the demand meanwhile a large portion of coins
desiring to leave the staking system would decrease the demand.