Add new param for price func and output to results.
This adds a new CLI option named 'pf' which allows the price ticket
calculation function to be chosen.  Currently, there are 4.  The
first one, named 'current', is the current algorithm used on mainnet as
of the date of this commit.  The other three are named '1', '2', and
'3', for the the first three proposals, respectively.

Also, it adds the chosen demand distribution function and ticket price
function to the output results along with a more user facing description
of each DDF and PF.