Use supply estimation in proposal 7.
This converts proposal 7 to use an estimate for the total supply instead
of the actual supply.  This is being done because even though the
simulator has access to the total supply, the end goal is that the final
concensus code only needs access to the most recent 2 intervals +
ticket maturity worth of block headers.  Calculating the actual supply
requires access to all previous headers.