Update proposal 7 with consensus-friendly int math.
This modifies proposal 7 to rearrange the equation and use integer math
instead of floating point math since floating point math can be
problematic across languages due to uncertainty in their respective
libs.  It also adds comments to further clarify the calculation and
rearranging of the equation.

Finally, while here, correct the previous retarget height in proposal 7
as well to use the block before the the previous retarget interval.
This is necessary because the stake difficulty must be calculated based
on existing blocks, and thus it is always calculated for the block after
a given block, which means the information for the previous retarget
interval must be calculated with the same conditions.